IErrorMask methods swapped from ushort indices to int (to match enums better)

+++ Loqui/Masks/IErrorMask.cs
@@ -10,7 +10,7 @@ namespace Loqui
     {
         Exception Overall { get; set; }
         List<string> Warnings { get; }
-        void SetNthException(ushort index, Exception ex);
-        void SetNthMask(ushort index, object maskObj);
+        void SetNthException(int index, Exception ex);
+        void SetNthMask(int index, object maskObj);
     }
 }
